Remove NonDetEff.hs and do everything in NonDet.hs.
This is a backwards-incompatible release, so this is now 0.3.0.0.
If we don't want to do this now, we can revert this change.
